display dogs and routes on frontend

+++ src/main/java/com/kamennova/doggies/dog/DogBreedRepository.java
@@ -2,5 +2,5 @@ package com.kamennova.doggies.dog;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DogBreedRepository extends JpaRepository<DogBreed, Integer> {
+public interface DogBreedRepository extends JpaRepository<DogBreed,Short> {
 }
